pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["imaanick <nickolaytrusov01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.35.70"
aiohttp = "^3.10.10"
fastapi = "^0.115.4"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
aiosqlite = "^0.20.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.13.3"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
aioresponses = "^0.7.6"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
